Extract highlightSubstring into SubstringHighlight component.
